neural-networks/speech-recognition/whisper-tiny-en/requirements.txt
--extra-index-url https://artifacts.luxonis.com/artifactory/luxonis-python-release-local/
depthai==3.0.0b1
git+https://github.com/luxonis/depthai-nodes.git@main#egg=depthai-nodes
opencv-python-headless~=4.10.0
numpy>=1.22
scipy==1.14.1
tqdm
openai-whisper